cleaned up CA failure conditions

+++ src_mega_reads/mega_reads_assemble_nomatch.sh
@@ -109,8 +109,8 @@ rm -f .rerun
 ###############removing redundant subreads or reducing the coverage by picking the longest reads##############################
 PB_SIZE=$(stat -c%s $PACBIO);
 if [ $B -lt 15 ];then
-echo "Detected nanopore data, we have to rename the reads";
 if [ ! -s "nanoporeRenamed.fa" ] ;then
+cho "Detected nanopore data, we have to rename the reads";
 awk 'BEGIN{n=0}{if($1 ~ /^>/){print $1"."n;n++}else{print $0}}' $PACBIO > nanoporeRenamed.fa;
 fi
 PACBIO1="nanoporeRenamed.fa";
@@ -346,46 +346,56 @@ cnsReuseUnitigs=1" > runCA.spec
 
 echo "Running assembly"
 if [ ! -e "${CA}/5-consensus/consensus.success" ]; then 
-#need to start from the beginning
-runCA -s runCA.spec consensus=pbutgcns -p genome -d $CA stopAfter=consensusAfterUnitigger $COORDS.1.frg $COORDS.1.mates.frg $SR_FRG $OTHER_FRG 1> $CA.log 2>&1 
-rm -rf $CA/5-consensus/*.success $CA/5-consensus/consensus.sh
-runCA -s runCA.spec -p genome -d $CA  stopAfter=consensusAfterUnitigger $COORDS.1.frg $SR_FRG $OTHER_FRG 1>> $CA.log 2>&1
+  #need to start from the beginning
+  runCA -s runCA.spec consensus=pbutgcns -p genome -d $CA stopAfter=consensusAfterUnitigger $COORDS.1.frg $COORDS.1.mates.frg $SR_FRG $OTHER_FRG 1> $CA.log 2>&1 
+  rm -rf $CA/5-consensus/*.success $CA/5-consensus/consensus.sh
+  runCA -s runCA.spec -p genome -d $CA  stopAfter=consensusAfterUnitigger $COORDS.1.frg $SR_FRG $OTHER_FRG 1>> $CA.log 2>&1
 fi
 
-#at athis point we assume that the unitig consensus is done
+#at athis point we check if the unitig consensus is done
 if [ ! -e "${CA}/5-consensus/consensus.success" ]; then
-echo "CA failure" && tail -n 40 $CA.log 
-exit;
+  echo "Assembly stopped or failed, see $CA.log" 
+  exit;
 fi
 
 if [ ! -e "${CA}/deduplicate.success" ]; then
 #here we remove overlaps to the reads in duplicate/redundant unitigs and then re-run the unitigger/consensus
-deduplicate_unitigs.sh $CA_PATH $CA genome $NUM_THREADS $OVL_MER $PLOIDY
+  deduplicate_unitigs.sh $CA_PATH $CA genome $NUM_THREADS $OVL_MER $PLOIDY
 fi
 
-#rerun CA on deduplicated overlapStore
-runCA -s runCA.spec consensus=pbutgcns -p genome -d $CA  stopAfter=consensusAfterUnitigger $COORDS.1.frg $COORDS.1.mates.frg $SR_FRG $OTHER_FRG 1>> $CA.log 2>&1
-rm -rf $CA/5-consensus/*.success $CA/5-consensus/consensus.sh
-runCA -s runCA.spec -p genome -d $CA  stopAfter=consensusAfterUnitigger cnsConcurrency=$(($NUM_THREADS/2+1)) $COORDS.1.frg $COORDS.1.mates.frg $SR_FRG $OTHER_FRG 1>> $CA.log 2>&1
 if [ ! -e "${CA}/5-consensus/consensus.success" ]; then
-echo "CA failure" && tail -n 40 $CA.log
-exit;
+  #after deduplicate we need to rebuild the unitigs, we rerun CA on deduplicated overlapStore
+  runCA -s runCA.spec consensus=pbutgcns -p genome -d $CA  stopAfter=consensusAfterUnitigger $COORDS.1.frg $COORDS.1.mates.frg $SR_FRG $OTHER_FRG 1>> $CA.log 2>&1
+  rm -rf $CA/5-consensus/*.success $CA/5-consensus/consensus.sh
+  runCA -s runCA.spec -p genome -d $CA  stopAfter=consensusAfterUnitigger cnsConcurrency=$(($NUM_THREADS/2+1)) $COORDS.1.frg $COORDS.1.mates.frg $SR_FRG $OTHER_FRG 1>> $CA.log 2>&1
+fi
+
+if [ ! -e "${CA}/5-consensus/consensus.success" ]; then
+  echo "Assembly stopped or failed, see $CA.log"
+  exit;
 fi
 
 #recompute astat if low pacbio coverage
 if [ $MCOVERAGE -le 5 ]; then
-if [ ! -e ${CA}/recompute_astat.success ];then
-recompute_astat_superreads_CA8.sh genome $CA $PE_AVG_READ_LENGTH $MASURCA_ASSEMBLY_WORK1_PATH/readPlacementsInSuperReads.final.read.superRead.offset.ori.txt  $SR_FRG
-fi
+  if [ ! -e ${CA}/recompute_astat.success ];then
+  recompute_astat_superreads_CA8.sh genome $CA $PE_AVG_READ_LENGTH $MASURCA_ASSEMBLY_WORK1_PATH/readPlacementsInSuperReads.final.read.superRead.offset.ori.txt  $SR_FRG
+  fi
 fi
 
 #we start from here if the scaffolder has been run or continue here  
 runCA -s runCA.spec consensus=pbutgcns -p genome -d $CA  stopAfter=consensusAfterScaffolder $COORDS.1.frg $COORDS.1.mates.frg $SR_FRG $OTHER_FRG 1>> $CA.log 2>&1
 rm -rf $CA/8-consensus/*.success $CA/8-consensus/consensus.sh
 runCA -s runCA.spec -p genome -d $CA  cnsConcurrency=$(($NUM_THREADS/2+1)) $COORDS.1.frg $COORDS.1.mates.frg $SR_FRG $OTHER_FRG 1>> $CA.log 2>&1 && \
-echo "Assembly complete, now cleaning up the scaffolds." 
-if [ ! -s $CA/dedup.genome.scf.fasta ];then
-deduplicate_contigs.sh $CA genome $NUM_THREADS $PLOIDY && echo "Final scaffold sequences are in $CA/dedup.genome.scf.fasta"
+echo "Assembly complete, now cleaning up the scaffolds." || echo "Assembly stopped or failed, see $CA.log"
+
+if [ -s $CA/9-terminator/genome.scf.fasta ];then
+  if [ ! -e $CA/filter_map.contigs.success ];then
+    deduplicate_contigs.sh $CA genome $NUM_THREADS $PLOIDY && echo "Assembly complete, final scaffold sequences are in $CA/dedup.genome.scf.fasta"
+  else
+    echo "Assembly complete, final scaffold sequences are in $CA/dedup.genome.scf.fasta"
+  fi
 else
-echo "Final scaffold sequences are in $CA/dedup.genome.scf.fasta"
+  echo "Assembly stopped or failed, see $CA.log"
 fi
+
+
